requirements.txt
accelerate
datasets
hydra-core
omegaconf
openai
transformers==4.37.2
wandb
torch
fire
trl
deepspeed
matplotlib
vllm
seaborn